+++ host/tests/L3_TestCases/dsCompositeIn/dsCompositeIn_test4_VerifySignal_Callback.py
@@ -23,15 +23,16 @@
 
 import os
 import sys
+from enum import Enum, auto
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
-sys.path.append(os.path.join(dir_path, "../"))
+sys.path.append(os.path.join(dir_path, "../../"))
 
 from dsCompositeInHelperClass import dsCompositeInHelperClass
 
-class dsCompositeIn_test3_ScaleAndVerify_Video(dsCompositeInHelperClass):
+class dsCompositeIn_test4_VerifySignal_Callback(dsCompositeInHelperClass):
     """
-    Test class to scale and verify the video in CompositeIn device
+    Test class to verify the signal change callback of CompositeIn device
 
     This class uses the `dscompositeInClass` to interact with the device's compositeIn ports,
     downloading necessary test assets, selecting
@@ -40,75 +41,77 @@ class dsCompositeIn_test3_ScaleAndVerify_Video(dsCompositeInHelperClass):
 
     def __init__(self):
         """
-        Initializes the test3_ScaleAndVerify_Video test .
+        Initializes the test4_VerifySignal_Callback test .
 
         Args:
             None.
         """
-        self.testName  = "test3_ScaleAndVerify_Video"
-        super().__init__(self.testName, '3')
+        self.testName  = "test4_VerifySignal_Callback"
+        super().__init__(self.testName, '4')
 
-    #TODO: Current version supports only manual verification.
-    def CheckDeviceStatusAndVerifyVideoScale(self, manual=False, port_type:str=0, videoscale:str=0):
+     #TODO: Current version supports only manual verification.
+    def CheckDeviceStatus(self, manual=False, connect=False, port_type:str=0):
         """
-        Verifies whether the particular port video scaled or not.
+        Verifies whether Composite Source Device connected or not
 
         Args:
+            port_type (str) : CompositeIn port
             manual (bool, optional): Manual verification (True: manual, False: other verification methods).
                                      Defaults to other verification methods
 
         Returns:
             bool
         """
-        if manual == True and videoscale == False:
-            return self.testUserResponse.getUserYN(f"Check CompositeIn source device connected&active on port {port_type} and press Y:")
-        elif manual == True and videoscale == True:
-            return self.testUserResponse.getUserYN(f"Check video scaled on {port_type} and press (Y/N)")
-        else:
+        if manual == True and connect == True:
+            return self.testUserResponse.getUserYN(f"Check if CompositeIn source is connected to {port_type} and press Y:")
+        elif manual == True and connect == False:
+            return self.testUserResponse.getUserYN(f"Check if CompositeIn source is disconnected to {port_type}/set to standby and press Y:")
+        else :
             #TODO: Add automation verification methods
             return False
 
     def testFunction(self):
         """
-        The main test function test scale video of dsCompositeIn device.
+        The main test function that verifies signal status of CompositeIn device.
 
         This function:
-        - Verifies video scaling .
+        - Verifies CompositeIn signal status through callbacks.
 
         Returns:
             bool: Final result of the test.
         """
 
-        self.log.testStart("test3_ScaleAndVerify_Video", '3')
+        self.log.testStart("test4_VerifySignal_Callback", '4')
 
         # Initialize the dsCompositeIn module
         self.testdsCompositeIn.initialise()
 
-        # x-coordiante, y-coordinate, width, height list
-        videoScale_argList = [[0,0,720,576], [500,500,500,500], [0,0,1920,1080]]
         # Loop through the supported CompositeIn ports
         for port in self.testdsCompositeIn.getSupportedPorts():
-            self.log.stepStart(f'Select {port} Port')
 
-            # Check the ComposteIn device connected to is active
+            # Check the CompositeIn device connected to is active
             portstr = f"dsCOMPOSITE_IN_PORT_{port}"
-            # Select the ComposteIn port
-            result = self.CheckDeviceStatusAndVerifyVideoScale(True,portstr,False)
-            self.log.stepResult(result,f'CompositeIn Device connected {result} on {portstr}')
-            self.testdsCompositeIn.selectPort(portstr)
+            for i in range(0, 2):
+                self.log.stepStart(f'Select {port} Port')
+                connect = True
+                if ( i == 1): connect = False
+                result = self.CheckDeviceStatus(True, connect, portstr)
+                self.log.stepResult(result,f'CompositeIn Device connected {result} on {portstr}')
 
-            # video scaling of ComposteIn port
-            for xvalue, yvalue, width, height in videoScale_argList:
-                self.log.stepStart(f'Scale video playing on compositeIn device')
-                self.testdsCompositeIn.scaleVideo(xvalue, yvalue, width, height)
-                result = self.CheckDeviceStatusAndVerifyVideoScale(True,portstr,True)
-                self.log.stepResult(result, f'CompositeIn Video Scale Verification {portstr} Port')
+                self.testdsCompositeIn.selectPort(portstr)
+                self.log.stepStart(f'Port Selcted {portstr}')
 
-        # Terminate dsCompositeIn Module
+                status = self.testdsCompositeIn.getSignalChangeCallbackStatus()
+                if status != None and portstr == status[0]:
+                    self.log.stepResult(True,f'Signal status {status[1]} found in Callback')
+                else:
+                    self.log.stepResult(False,f'Signal status not found in Callback')
+
+        # Terminate testdsCompositeIn Module
         self.testdsCompositeIn.terminate()
 
         return result
 
 if __name__ == '__main__':
-    test = dsCompositeIn_test3_ScaleAndVerify_Video()
+    test = dsCompositeIn_test4_VerifySignal_Callback()
     test.run(False)
